mv business logic to view model class
